Avoid leaking Go pointers, and free all C pointers
"Free all C pointers" is somewhat of a misnomer, as holes will leak if
you run NewTriangulateIO yourself, because of a double free issue I
haven't properly fixed. There are probably tricks one can use to ensure
data isn't double free'd. Reference counters are probably the easiest
to use, but it feels a bit ugly to wrap that atop of this fix (and also
this library).

Also, there are fields here that I haven't experimented with, so it's
quite likely that using those may result in double frees.